requirements.txt
Flask
xhtml2pdf
openai
python-dotenv